requirements.txt
flask>=3.0.0
pynput>=1.7.6
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0
openai>=1.0.0
